+++ persistent/ChangeLog.md
@@ -36,6 +36,9 @@
     ```haskell
     deriving stock instance Show (Unique User)
     ```
+
+* [#1131](https://github.com/yesodweb/persistent/pull/1131)
+  * Add an `exists` function to the `PersistQueryRead` type class.
 * [#1117](https://github.com/yesodweb/persistent/issues/1117)
   * Allow parsing UTCTimes from sqlite with the format "%F %T%Q" as well, instead of only "%FT%T%Q".
 * [#1140](https://github.com/yesodweb/persistent/pull/1140)